+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # ImageOptimize Changelog
 
+## 4.0.7 - 2024.04.04
+### Added
+* Added the ability to pass in a config array to `.imgTag()`, `.pictureTag()` and `.linkPreloadTag()`
+
+### Changed
+* Changed `.loading()` → `.loadingStrategy()`, `.artDirection()` → `addSourceFrom()`
+
 ## 4.0.6 - 2024.04.02
 ### Added
 * Added a `--queue` option to the `image-optimize/optimize/create` console command if you want to defer the image generation to be run via queue job, rather than immediately via the console command